DRY-up *PendingTransactionSorter tests
